requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

dnspython=2.2.1=py311h06a4308_0
flask=2.2.2=py311h06a4308_0
pydantic=1.10.2=py311h5eee18b_0
pymongo=3.12.0=py311h6a678d5_0
python=3.11.3=h7a1cb2a_0
python-dotenv=0.21.0=py311h06a4308_0
